requirements.txt
python-binance
pandas
numpy
requests